[Rename] #31 - Rename main component
