+++ presentation/src/main/java/com/woowahan/banchan/ui/root/RootTabPagerAdapter.kt
@@ -1,10 +1,10 @@
-package com.woowahan.banchan.ui.main
+package com.woowahan.banchan.ui.root
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 
-class MainTabPagerAdapter(
+class RootTabPagerAdapter(
     fragmentActivity: FragmentActivity,
     private val fragmentList: List<Pair<String, Fragment>>
 ): FragmentStateAdapter(fragmentActivity) {
