+++ presentation/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.main.MainActivity">
+        tools:context=".ui.root.RootActivity">
 
         <com.google.android.material.appbar.AppBarLayout
             app:layout_constraintEnd_toEndOf="parent"
